Run integration tests only when that package has been changed (#45)
For instance, GCP integration tests will only run when any file inside
`pkg/gcp` has changes. This saves time on running expensive integration
tests, especially on PRs that don't touch these packages at all.

This is not done for unit tests as unit tests are fast. However, this
can be very easily extended to unit tests in the future if need be. All
we would need to do is move up the actions added in this PR to before
the unit tests run.