[GCP] Add connectivity tests to integration tests (#51)
Added two connectivity tests that ensure that the VMs created during
integration tests can ping each other.

Had to make some changes to allow for "optional" ports as ICMP doesn't
have the notion of ports. Designated -1 to be an optional, which **user
has to specify -1 explicitly**. This is because GCP allows ports to be 0
for TCP/UDP. GRPC will automatically set the port to 0 if it's not
specified.