[GCP] Remove specifying necessary clients for unit tests (#89)
Fixes #71.

Regarding closing the clients (and the http server), there's no
difference in functionality as all of them will get closed once the test
is done. It probably is good practice akin to closing file descriptors.
~~However, considering it saves us from remembering to add a new GCP
client to a teardown method and calling teardown, it seems okay to
forego this "good practice".~~ We will actually keep them as good
practice.