Implement basic CLI skeleton
Part-of: #111

This change implements the skeleton of the CLI based on the description in the issue. The overall structure matches the structure I've used on other projects successfully.

The main change here besides adding the new structure is to split server and client functionality. We already build two CLI binaries, but one of them was not being used.

- `inv` is for client functionality (creating tags, rules)
- `invd` is for server functionality (running the frontend, tag server, etc.)