Added Namespaces to Existing Plugins (#100)
In this PR:

* Created namespaces (#23 )
* Namespaces are set on the frontend (local state) and scopes all
requests to cloud plugins
* Cloud plugins use the namespace to separate resources into different
vpcs/vnets and refuse to interact with resources if they are not in the
set namespace
* Changed how each server is initialized
* Recently, the `Setup` functions were changed to be async, but this
does not give the intended effect for running the servers locally via
the CLI. Now the integration tests run the `Setup` function as a
goroutine.
* The plugins now accept the frontend address as a parameter to their
`Setup`. This removes the need for the `FrontendServerAddr` global
variable.
  * Resolved unused ID field issue on the frontend (#48 )
  
 Future work:
* Decide how to connect namespaces. As of now, the central controller
gives out overlapping address spaces to different namespaces. We should
decide if this is the right design and if not, decide how to connect
different namespaces (eg, VPN ? VNET peering?)

---------

Co-authored-by: Sean Kim <kimsean@google.com>