[GCP][Azure] Add get checks for multicloud resources (#110)
We used to rely on idempotence, but the seemingly indifferent updates
take a while and actually triggers an update on the Azure side (unclear
whether GCP does this). This tells us that the "updates" we apply, while
they don't affect correctness, they are different from whatever Azure is
holding. These updates also cause the gateways to lose connection
briefly due to the update, but they come back and don't affect
correctness.

We should have a better fix for this in the future and see how we can
actually rely on idempotence since that will save us a get call, but for
now, get checks have been added. For GCP, I only added get checks for
patch requests. The insert requests already fail if the resource already
exists, so I assumed there wouldn't be much of a difference adding a get
before insert.

Fixes #107.