[Azure] Add hub and spoke model for VPN gateway (#126)
Fixes #116.

This PR makes a standalone vnet for the VPN gateway, which peers to all
other vnets.

![image](https://github.com/NetSys/invisinets/assets/12043327/0f73bd6c-b98c-483e-928a-fbfe0ae63101)

I've amended the multicloud test to create a second VM in a different
region and ensure connectivity between that second VM and GCP VM is
valid.

Some technical details to note
- The VPN gateway vnet is created on the very first resource creation
(i.e. along with the first vnet).
- On every resource creation, we create a peering between the vnet the
resource resides in and the VPN gateway vnet if it doesn't already
exist.
- Azure doesn't allow gateway transit peering relationships to be
established unless the VPN gateway vnet has a VPN gateway provisioned.
While we don't want provision a VPN gateway before any multicloud permit
list rules are added (due to costs), we still create the vnet and
establish peerings which is free. Later on, when we do create the VPN
gateway, we will retoractively modify the peerings to establish the
gateway transit peering relationship.
- If the VPN gateway already exists, then we will simply create peerings
with the gateway transit peering relationship enabled.