Remove address space for connecting VPN gateways (#132)
Originally, I thought we needed to specify the local network address
space when creating the local network gateway. To give a concrete
example, if GCP VPC had an address space of 10.0.0.0/16, then I thought
I needed to put the 10.0.0.0/16 as a local network address prefix
property when creating the local network gateway in Azure. But after
reading the description (from the Azure portal of creating local network
gateway) which states

> One or more IP address ranges (in CIDR notation) that define your
local network's address space. For example: 192.168.0.0/16. If you plan
to use this local network gateway in a BGP-enabled connection, then the
minimum prefix you need to declare is the host address of your BGP Peer
IP address on your VPN device.

removing the address space field in our `CreateVpnConnectionsRequest` is
the way to go since both Azure and GCP doesn't need it.