Fix resource naming conventions for multicloud to incorporate namespaces (#135)
Adds namespaces to the prefix of the resource name when required. To be
precise, namespaces are incorporated into the resource name if another
namespace in the same resource group (for Azure) or same project (for
GCP) could cause a collision (e.g. vpn gateways, local network gateways,
cloud routers, and most other things). Things that don't incorporate
namespaces are "nested" inside the resource. These would be things like
the BGP peerings in a GCP cloud router which have a unique name per
cloud router.

Now, an argument could be made that some resources could be shared (e.g.
local network gateways in Azure if both namespaces want to reference the
same GCP gateway). That's been left off for now, ensuring a clean and
complete separation between namespaces. Perhaps in the future, we can
look at it again and see if we can make some optimizations to reduce
resource costs.