[GCP][Azure] Add support for cross-namespaces intra-cloud peerings (#162)
I added peering functionality for cross-namespaces intra-cloud
connectivity. When connecting across namespaces within the same cloud, a
peering functionality ([VPC network
peering](https://cloud.google.com/vpc/docs/vpc-peering) for GCP,
[virtual network
peering](https://learn.microsoft.com/en-us/azure/virtual-network/virtual-network-peering-overview)
for Azure) will be used.

**Mesh Problem**
If there are too many cross-namespaces connections, this can create a
messy mesh of different peerings. @smcclure20 and I thought it can be
good to give users an option to instead establish VPN gateway
connections for cross-namespace intra-cloud connections that require a
lot of peerings. This would go through the existing
`orchestrator.ConnectClouds` infrastructure where the namespaces would
operate as if they are part of separate clouds. We can implement this in
a future PR if we see this as a good design choice.

**Problems with `AzureSDKHandler`**
I think it's worth revisiting #69 #70 again in a more serious way. I've
had to put in a lot of hacks to make this work. The major change was to
pass to remove `azureHandler` (of type `AzureSDKHandler`) from the
`azurePluginServer`. This was necessary because `azureHandler` might be
changed during the execution of an Azure RPC request if that Azure RPC
request calls the orchestrator which then invokes another Azure RPC
request (e.g., `azure_plugin.CreateResource` ->
`orchestrator.FindUnusedAddressSpace` ->
`azure_plugin.GetUsedAddressSpaces`). This was not much of a hack, but
having to deal with changes in our unit testing infrastructure and the
mocking scheme was very hacky. Overall, I don't see a need for a
`AzureSDKHandler` interface and should just create methods that take in
subscription IDs and resource group names like how GCP does it with
project names.