[GCP] Fix GitHub Actions authentication (#182)
Seems like something went wrong with GCP test permissions when we
transferred ownership of the repo to a new org. GCP authentication was
fine but running integration tests failed due to permission issues.

We may have to do this again after we find a new name for this project
(#180). Keeping in mind #181 as well.