protoc-gen-go not being found on OS_ARCH pairs other than linux_amd64 (#199)
`go install` will install binaries for the specified `GOOS` and
`GOARCH`. While we specify different `GOOS` and `GOARCH` according to
the build matrix, the GitHub runners that we are currently working with
are linux amd64 machines (according to `go env GOHOSTOS` and `go env
GOHOSTARCH`). As a result, `go install` put the Go GRPC libraries into
`~/go/bin/${GOOS}_${GOARCH}` instead of directly into `~/go/bin`.
`$PATH` only included `~/go/bin`, which is why the failed runs in #188
were saying `protoc-gen-go` couldn't be found.

I tried adding `~/go/bin/${GOOS}_${GOARCH}` to `$PATH`, but that didn't
turn out to be correct since I was trying to run binaries that were
compiled for a different architecture than the host. The correct thing
to do was install the Go GRPC libraries for the host OS/architecture,
which is why the `go install` are prepended with temporarily changed
`GOOS` and `GOARCH` variables.

Also, the `go get`s are unnecessary.