[Docs] Use Sphinx docs hosted on GitHub Pages (#186)
I didn't change much of the content as we can do that in subsequent PRs.
This is mainly just a setup for Sphinx and GitHub Pages. Main change is
using RST files instead of MD files as Sphinx natively supports RST. The
syntax is slightly different but it should be much better and forward
looking for generating much more complex and involved documentation.

~~Still considering [Read the Docs](https://readthedocs.com/), but it
would come in a future PR if we decide it's necessary. A full comparison
between Read the Docs and GitHub Pages:
https://about.readthedocs.com/comparisons/github-pages/. One of the main
advantages I can see is the deploy previews which we get to preview the
docs before merging a PR. With GitHub Pages, we can't do that.~~ I've
enabled previews using
https://github.com/marketplace/actions/deploy-pr-preview. It should be
available at https://invisinets.github.io/invisinets/pr-preview/pr-186/