[Azure][GCP] Implement Support for Managed K8s Clusters (#175)
In this PR:
* Generalizes assumptions in existing code that a resource is an
instance/VM
* Implements support for k8s clusters in Azure and GCP
* In Azure, managed clusters are given their own subnet and all
invisinets permit list rules will be tied to that subnet (since the
cluster does not have a single NIC that can have its own NSG)
* In GCP, the cluster ID is used to create a firewall tag and all the
rest works exactly as instances do

In this
[doc](https://docs.google.com/document/d/1wTgjEGTMbBL5nPSjhLDnu7_3FEw0afIUmCN9EEgemFk/edit?usp=sharing)
is my working understanding of the relevant pieces of k8s networking and
what that means for Invisinets. This doc can serve as a high-level
overview of what the PR is attempting to do.