Add deployment and resource name to `CreateResource` (#211)
Orchestrator will now pass in the deployment and resource name to a
`CreateResource` request to a cloud plugin. This is so that the user
does not have to specify this type of information in the JSON provided.
If the user provides any of this info in the JSON, it will simply be
overwritten with the deployment and resource name that's passed in
through the method.